Move leaf-gp into optimizer folders
