+++ alfalfa/optimizer/__init__.py
@@ -1,0 +1,2 @@
+from .gbm_model import GbmModel
+from .proposals import propose
